+++ scripts/module.js
@@ -56,14 +56,14 @@ Hooks.once('ready', async function() {
 			oldVersion = 3;
 		}
 		await setupJournalEntry();
-		await tashaSummon.setupFolder();
+		if (game.settings.get('chris-premades', 'Tasha Actors')) await tashaSummon.setupFolder();
 		if (game.settings.get('itemacro', 'charsheet')) ui.notifications.error('Chris\'s Premades & Midi-Qol requires "Character Sheet Hook" in Item Macro\'s module settings to be turned off!');
 		Hooks.on('getItemSheetHeaderButtons', createHeaderButton);
 		if (game.modules.get('ddb-importer')?.active) Hooks.on('getActorSheet5eHeaderButtons', createActorHeaderButton);
 		game.settings.set('chris-premades', 'LastGM', game.user.id);
-		if (game.settings.get('chris-premades', 'Combat Listener') && game.user.isGM) Hooks.on('updateCombat', combatUpdate);
-		if (game.settings.get('chris-premades', 'Movement Listener') && game.user.isGM) Hooks.on('updateToken', tokenMoved);
-		if (game.settings.get('chris-premades', 'Effect Auras') && game.user.isGM) {
+		if (game.settings.get('chris-premades', 'Combat Listener')) Hooks.on('updateCombat', combatUpdate);
+		if (game.settings.get('chris-premades', 'Movement Listener')) Hooks.on('updateToken', tokenMoved);
+		if (game.settings.get('chris-premades', 'Effect Auras')) {
 			Hooks.on('preUpdateActor', effectAuraHooks.preActorUpdate);
 			Hooks.on('updateActor', effectAuraHooks.actorUpdate);
 			Hooks.on('canvasReady', effectAuraHooks.canvasReady);